feat: Rename asset module to process module
